[UPSTREAM][MIPS] Disassemble the 0xefefefef trap padding used by LLD as `trap_ri`
Previously using llvm-objdump would just print a lot of 0xefefefef unknown
instructions. If we disassemble this as `trap_ri` it should be obvious that
it is a explicit trap padding rather than some instruction that is not
available because the wrong -triple was passed to llvm-objdump.

I am not sure if there `trap_ri` is the best mnemonic to use here, I just
decided to go for that since it will always raise a reserved instruction
exception.