Make &(somecap->field) a capability.
We handle a similar case of an l-value to r-value cast in the back end.
We were somewhat inconsistent about this or structures, because
`somecap->field` would do a capability-relative load, but
`*&somecap->field` would not.

We should probably do the same thing for arrays, so that &somecap[x]
evaluates to a capability as well.

Fixes #180