There is no need to call getTargetAddressSpace in Sema
My last change to getTargetAddressSpace made cheri-warn-provenance-lost.c
fail. This turns out is due to getPointerRange() returning 128 for CHERI!
It is also getting the address space of the pointee type and not the
pointer so was only working by chance.
This fixes the test by using ASTContext::getIntRange() instead.

Also update the test to also check the hybrid ABI